#3-13. Add board update, delete / Add icon in board_list

+++ daham/views.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 
+from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 
 now = datetime.now()
@@ -55,6 +56,9 @@ def detail(request, board_id):
 @login_required(login_url='common:login')
 def board_update(request, board_id):
     board = get_object_or_404(Board, pk=board_id)
+    if request.user != board.user:
+        messages.error(request, '수정권한이 없습니다.')
+        return redirect('daham:detail', board_id=board.id)
 
     if request.method == "POST":
         form = BoardForm(request.POST, instance=board)  # 기존 값을 폼에 채우기
@@ -73,6 +77,9 @@ def board_update(request, board_id):
 @login_required(login_url='common:login')
 def board_delete(request, board_id):
     board = get_object_or_404(Board, pk=board_id)
+    if request.user != board.user:
+        messages.error(request, '삭제권한이 없습니다.')
+        return redirect('daham:detail', board_id=board.id)
     board.delete()
     return redirect('daham:board')
